+++ CHANGES.rst
@@ -104,6 +104,8 @@ FIXED:
 * FIXED: Some tests related to python-3.11
 * FIXED: Some tests related to python-3.9
 * FIXED: active-tag logic if multiple tags with same category exists.
+* pull  #1249: Fixed having one event loop per async step for python >= 3.10 (provided-by: jeteve; partial)
+* issue #1233; RuntimeError: Event loop is closed" in asyncio streams (submitted by: rdpoor)
 * issue #1231: TypeConverter error causes error in "pretty" formatter (provided-by: rittneje)
 * issue #1202: behave fails with pyproject.toml without any [tool] sections (provided by: con-fah)
 * issue #1177: Bad type-converter pattern: MatchWithError is turned into AmbiguousStep (submitted by: omrischwarz)